+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ENHANCEMENTS:
 * mdb: add `maintenance_window` section in `yandex_mdb_mongodb_cluster` resource and data source
 
+BUG FIXES:
+* mdb: fixed some errors in implementation of Kafka topics modification
+
 ## 0.59.0 (June 6, 2021)
 FEATURES:
 * **New Data Source:** `yandex_alb_http_router`